Move gascalibration to tools and add submodule
