+++ .gitignore
@@ -6,8 +6,7 @@ wasp.log
 /evmemulator
 /schema
 /submsg
-/vendor
-waspdb/
+waspdb**/
 cluster-data/
 dist/
 .idea/
@@ -25,17 +24,14 @@ packages/gpa/**/states/
 packages/gpa/**/*.toolbox/
 wasp-cli.json
 wasp-config-template.json
-**/pkg/
-**/target/
+contracts/**/pkg/
+contracts/**/target/
 .DS_Store
-./tools/wasp-cli/wasp-cli
 .docusaurus
 **/node_modules/
-goshimmer.log
 /*.sh
 wal/
 snap-cli
 *.trie
 *.snapshot
 snapshot
-gascalibration
