+++ tools/wasp-cli/chain/alias.go
@@ -5,8 +5,8 @@ import (
 	"github.com/spf13/viper"
 
 	"github.com/iotaledger/wasp/packages/isc"
-	"github.com/iotaledger/wasp/packages/log"
 	"github.com/iotaledger/wasp/tools/wasp-cli/config"
+	"github.com/iotaledger/wasp/tools/wasp-cli/log"
 )
 
 var chainAlias string
