+++ tools/wasp-cli/chain/cmd.go
@@ -3,7 +3,7 @@ package chain
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/iotaledger/wasp/packages/log"
+	"github.com/iotaledger/wasp/tools/wasp-cli/log"
 )
 
 var chainCmd = &cobra.Command{
