+++ tools/wasp-cli/chain/events.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/iotaledger/wasp/packages/isc"
 	"github.com/iotaledger/wasp/packages/kv/dict"
-	"github.com/iotaledger/wasp/packages/log"
 	"github.com/iotaledger/wasp/packages/vm/core/blocklog"
+	"github.com/iotaledger/wasp/tools/wasp-cli/log"
 )
 
 var eventsCmd = &cobra.Command{
