+++ tools/wasp-cli/chain/listcontracts.go
@@ -4,8 +4,8 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/iotaledger/wasp/packages/kv/collections"
-	"github.com/iotaledger/wasp/packages/log"
 	"github.com/iotaledger/wasp/packages/vm/core/root"
+	"github.com/iotaledger/wasp/tools/wasp-cli/log"
 )
 
 var listContractsCmd = &cobra.Command{
