+++ tools/wasp-cli/peering/cmd.go
@@ -6,7 +6,7 @@ package peering
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/iotaledger/wasp/packages/log"
+	"github.com/iotaledger/wasp/tools/wasp-cli/log"
 )
 
 var peeringCmd = &cobra.Command{
