+++ tools/wasp-cli/peering/distrust.go
@@ -6,9 +6,9 @@ package peering
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/iotaledger/wasp/packages/log"
 	"github.com/iotaledger/wasp/packages/peering"
 	"github.com/iotaledger/wasp/tools/wasp-cli/config"
+	"github.com/iotaledger/wasp/tools/wasp-cli/log"
 )
 
 var distrustCmd = &cobra.Command{
