+++ tools/wasp-cli/peering/listtrusted.go
@@ -6,8 +6,8 @@ package peering
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/iotaledger/wasp/packages/log"
 	"github.com/iotaledger/wasp/tools/wasp-cli/config"
+	"github.com/iotaledger/wasp/tools/wasp-cli/log"
 )
 
 var listTrustedCmd = &cobra.Command{
