+++ tools/wasp-cli/util/token_parsing.go
@@ -7,7 +7,7 @@ import (
 	"github.com/ethereum/go-ethereum/common/hexutil"
 
 	"github.com/iotaledger/wasp/packages/isc"
-	"github.com/iotaledger/wasp/packages/log"
+	"github.com/iotaledger/wasp/tools/wasp-cli/log"
 )
 
 const BaseTokenStr = "base"
